// apps/desktop/src/generated/query/ChangeOriginDto.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.

export type ChangeOriginDto = { "type": "sync_run", sync_run_id: string, } | { "type": "binding", binding_id: string, } | { "type": "inference", rule_version: string, input_resource_version_ids: Array<string>, input_relation_version_ids: Array<string>, };