// apps/desktop/src/generated/query/TimelineItemDto.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { ChangeDto } from "./ChangeDto";
import type { TimelineVersionLinkDto } from "./TimelineVersionLinkDto";

export type TimelineItemDto = { change: ChangeDto, version_links: Array<TimelineVersionLinkDto>, };