// apps/desktop/src/generated/query/TimelinePageDto.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { PageInfo } from "./PageInfo";
import type { SnapshotMetadata } from "./SnapshotMetadata";
import type { TimelineGroupDto } from "./TimelineGroupDto";

export type TimelinePageDto = { metadata: SnapshotMetadata, groups: Array<TimelineGroupDto>, page_info: PageInfo, };